Fix CI native build and extract SSH auth callbacks to stay under lint limit
- Add cpu-features to pnpm.onlyBuiltDependencies for ssh2 native module
- Add build-essential to CI workflow for node-gyp compilation
- Extract SSH auth IPC callbacks into ssh-auth-helpers.ts (ssh.ts 388→297 lines)
