+++ .github/workflows/pr.yml
@@ -16,6 +16,9 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v4
 
+      - name: Install native build tools
+        run: sudo apt-get update && sudo apt-get install -y build-essential python3
+
       - name: Setup Node.js
         uses: actions/setup-node@v4
         with:
